+++ src/elements/images.rs
@@ -1,4 +1,5 @@
 // SPDX-FileCopyrightText: 2020 Alexander Dean-Kennedy <dstar@slackless.com>
+// SPDX-FileCopyrightText: 2021 Robin Krahl <robin.krahl@ireas.org>
 // SPDX-License-Identifier: Apache-2.0 or MIT
 
 //! Image support for genpdf-rs.
@@ -7,7 +8,7 @@ use std::path;
 
 use image::GenericImageView;
 
-use crate::error::{Context as _, Error};
+use crate::error::{Context as _, Error, ErrorKind};
 use crate::{render, style};
 use crate::{Alignment, Context, Element, Mm, Position, RenderResult, Rotation, Scale, Size};
 
@@ -20,7 +21,8 @@ use crate::{Alignment, Context, Element, Mm, Position, RenderResult, Rotation, S
 /// # Supported Formats
 ///
 /// All formats supported by the [`image`][] should be supported by this crate.  The BMP, JPEG and
-/// PNG formats are well tested and known to work.
+/// PNG formats are well tested and known to work.  Yet it is currently not possible to render
+/// images with transparency, see [`printpdf` issue #98][].
 ///
 /// Note that only the GIF, JPEG, PNG, PNM, TIFF and BMP formats are enabled by default.  If you
 /// want to use other formats, you have to add the `image` crate as a dependency and activate the
@@ -39,6 +41,7 @@ use crate::{Alignment, Context, Element, Mm, Position, RenderResult, Rotation, S
 ///
 /// [`image`]: https://lib.rs/crates/image
 /// [`printpdf::Image`]: https://docs.rs/printpdf/latest/printpdf/types/plugins/graphics/two_dimensional/image/struct.Image.html
+/// [`printpdf` issue #98]: https://github.com/fschutt/printpdf/issues/98
 #[derive(Clone)]
 pub struct Image {
     data: image::DynamicImage,
@@ -63,14 +66,21 @@ pub struct Image {
 
 impl Image {
     /// Creates a new image from an already loaded image.
-    pub fn from_dynamic_image(data: image::DynamicImage) -> Self {
-        Image {
-            data,
-            alignment: Alignment::default(),
-            position: None,
-            scale: Scale::default(),
-            rotation: Rotation::default(),
-            dpi: None,
+    pub fn from_dynamic_image(data: image::DynamicImage) -> Result<Self, Error> {
+        if data.color().has_alpha() {
+            Err(Error::new(
+                "Images with an alpha channel are not supported",
+                ErrorKind::InvalidData,
+            ))
+        } else {
+            Ok(Image {
+                data,
+                alignment: Alignment::default(),
+                position: None,
+                scale: Scale::default(),
+                rotation: Rotation::default(),
+                dpi: None,
+            })
         }
     }
 
@@ -85,7 +95,7 @@ impl Image {
             .context("Could not determine image format")?
             .decode()
             .context("Could not decode image")?;
-        Ok(Self::from_dynamic_image(image))
+        Self::from_dynamic_image(image)
     }
 
     /// Creates a new image from the given reader.
@@ -225,48 +235,283 @@ impl Element for Image {
     }
 }
 
-/// Given the Size of a box (width/height), compute the bounding-box size when
-/// rotated some degrees and where the "minimum" corner is (which should be the
-/// new origin/offset). Note, this is not very optimized.
+/// Given the Size of a box (width/height), compute the bounding-box size and offset when
+/// rotated some degrees.  The offset is the distance from the top-left corner of the bounding box
+/// to the (originally) lower-left corner of the image.
 #[allow(clippy::manual_range_contains)]
 fn bounding_box_offset_and_size(rotation: &Rotation, size: &Size) -> (Position, Size) {
-    let theta = rotation.degrees.to_radians();
-    let (ct, st) = (theta.cos(), theta.sin());
-    let (w, h): (f64, f64) = (size.width.into(), size.height.into());
-    match rotation.degrees {
-        d if d > 0.0 && d <= 90.0 => {
-            let alpha = 180.0 - (rotation.degrees + 90.0);
-            let ca = alpha.to_radians().cos();
-            let (hct, wct) = (h * ct, w * ct);
-            let (hst, wst) = (h * st, w * st);
-            let (bb_w, bb_h) = (hst + wct, wst + hct);
-            (Position::new(h * ca, bb_h), Size::new(bb_w, bb_h))
+    // alpha = rotation, beta = 90 - rotation
+    let alpha = rotation.degrees.to_radians();
+    let beta = (90.0 - rotation.degrees).to_radians();
+
+    // s* = sin of *
+    let sa = alpha.sin();
+    let sb = beta.sin();
+
+    // Bounding box calculation, based on
+    // https://math.stackexchange.com/questions/1628657/dimensions-of-a-rectangle-containing-a-rotated-rectangle
+    let width = (size.width.0 * sb).abs() + (size.height.0 * sa).abs();
+    let height = (size.height.0 * sb).abs() + (size.width.0 * sa).abs();
+    let bb_size = Size::new(width, height);
+
+    // Offset calculation -- to follow the calculations, consider the rotated rectangles, their
+    // bounding boxes and the triangles between them
+    let bb_position = if rotation.degrees < -180.0 {
+        unreachable!(
+            "Rotations must be in the range -180.0..=180.0, but got: {}",
+            rotation.degrees
+        );
+    } else if rotation.degrees <= -90.0 {
+        Position::new(size.width.0 * alpha.cos().abs(), 0)
+    } else if rotation.degrees <= 0.0 {
+        Position::new(0, size.height.0 * alpha.cos())
+    } else if rotation.degrees <= 90.0 {
+        Position::new(size.height.0 * beta.cos(), bb_size.height.0)
+    } else if rotation.degrees <= 180.0 {
+        Position::new(bb_size.width.0, size.width.0 * beta.cos())
+    } else {
+        unreachable!(
+            "Rotations must be in the range -180.0..=180.0, but got: {}",
+            rotation.degrees
+        );
+    };
+
+    (bb_position, bb_size)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::bounding_box_offset_and_size;
+    use crate::{Position, Rotation, Size};
+    use float_cmp::approx_eq;
+
+    macro_rules! assert_approx_eq {
+        ($typ:ty, $lhs:expr, $rhs:expr) => {
+            let left = $lhs;
+            let right = $rhs;
+            assert!(
+                approx_eq!($typ, left, right, epsilon = 100.0 * f64::EPSILON, ulps = 10),
+                "assertion failed: `(left approx_eq right)`
+  left: `{:?}`,
+ right: `{:?}`",
+                left,
+                right
+            );
+        };
+    }
+
+    fn test_position(size: Size, rotation: f64, position: Position) {
+        println!("rotation = {}", rotation);
+        let rotation = Rotation::from(rotation);
+        assert_approx_eq!(
+            Position,
+            position,
+            bounding_box_offset_and_size(&rotation, &size).0
+        );
+    }
+
+    #[test]
+    fn test_bounding_box_size_square_0_deg() {
+        let size = Size::new(100, 100);
+        for rotation in &[-180.0, -90.0, 0.0, 90.0, 180.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(Size, size, bounding_box_offset_and_size(&rotation, &size).1);
         }
-        d if d > 90.0 && d <= 180.0 => {
-            let alpha = (rotation.degrees - 90.0).to_radians();
-            let (ca, sa) = (alpha.cos(), alpha.sin());
-            let (bb_w, bb_h) = (w * sa + h * ca, w * ca + h * sa);
-            (Position::new(bb_w, w * ca), Size::new(bb_w, bb_h))
+    }
+
+    #[test]
+    fn test_bounding_box_size_square_30_deg() {
+        let size = Size::new(100, 100);
+        let bb_width = (60.0f64.to_radians().sin() + 30.0f64.to_radians().sin()) * size.width.0;
+        let bb_size = Size::new(bb_width, bb_width);
+        for rotation in &[-150.0, -120.0, -30.0, -60.0, 30.0, 60.0, 120.0, 150.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(
+                Size,
+                bb_size,
+                bounding_box_offset_and_size(&rotation, &size).1
+            );
         }
-        d if d < 0.0 && d > -90.0 => {
-            let (hct, wct) = (h * ct, w * ct);
-            let (hst, wst) = (h * st, w * st);
-            let (bb_w, bb_h) = (hst + wct, hct + wst);
-            (Position::new(0, hct), Size::new(bb_w, bb_h))
+    }
+
+    #[test]
+    fn test_bounding_box_size_square_45_deg() {
+        let size = Size::new(100, 100);
+        let bb_width = (2.0f64 * size.width.0.powf(2.0)).sqrt();
+        let bb_size = Size::new(bb_width, bb_width);
+        for rotation in &[-135.0, -45.0, 45.0, 135.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(
+                Size,
+                bb_size,
+                bounding_box_offset_and_size(&rotation, &size).1
+            );
         }
-        d if d <= -90.0 && d >= -180.0 => {
-            let alpha = (180.0 + rotation.degrees).to_radians();
-            let (ca, sa) = (alpha.cos(), alpha.sin());
-            let (bb_w, bb_h) = (h * sa + w * ca, h * ca + w * sa);
-            (Position::new(w * ca, 0), Size::new(bb_w, bb_h))
+    }
+
+    #[test]
+    fn test_bounding_box_position_square_30_deg() {
+        let size = Size::new(100, 100);
+        let bb_width =
+            30.0f64.to_radians().sin() * size.width.0 + 60.0f64.to_radians().sin() * size.height.0;
+
+        let w30 = 30.0f64.to_radians().cos() * size.width.0;
+        let w60 = 60.0f64.to_radians().cos() * size.width.0;
+
+        test_position(size, -150.0, Position::new(w30, 0));
+        test_position(size, -120.0, Position::new(w60, 0));
+        test_position(size, -60.0, Position::new(0, w60));
+        test_position(size, -30.0, Position::new(0, w30));
+        test_position(size, 30.0, Position::new(w60, bb_width));
+        test_position(size, 60.0, Position::new(w30, bb_width));
+        test_position(size, 120.0, Position::new(bb_width, bb_width - w60));
+        test_position(size, 150.0, Position::new(bb_width, bb_width - w30));
+    }
+
+    #[test]
+    fn test_bounding_box_position_square_45_deg() {
+        let size = Size::new(100, 100);
+        let bb_width = (2.0f64 * size.width.0.powf(2.0)).sqrt();
+
+        test_position(size, -135.0, Position::new(bb_width / 2.0, 0));
+        test_position(size, -45.0, Position::new(0, bb_width / 2.0));
+        test_position(size, 45.0, Position::new(bb_width / 2.0, bb_width));
+        test_position(size, 135.0, Position::new(bb_width, bb_width / 2.0));
+    }
+
+    #[test]
+    fn test_bounding_box_position_square_90_deg() {
+        let size = Size::new(100, 100);
+        test_position(size, -180.0, Position::new(100, 0));
+        test_position(size, -90.0, Position::new(0, 0));
+        test_position(size, 0.0, Position::new(0, 100));
+        test_position(size, 90.0, Position::new(100, 100));
+        test_position(size, 180.0, Position::new(100, 0));
+    }
+
+    #[test]
+    fn test_bounding_box_size_rectangle_0_deg() {
+        let size = Size::new(200, 100);
+        for rotation in &[-180.0, 0.0, 180.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(Size, size, bounding_box_offset_and_size(&rotation, &size).1);
         }
-        _ =>
-        // This section is only for degrees == 0.0, but I use the default match due to:
-        //       https://github.com/rust-lang/rust/issues/41620
-        // Rotation's degrees should be restricted to [-180,180] so these
-        // ranges should be complete.
-        {
-            (Position::new(0, h), *size)
+    }
+
+    #[test]
+    fn test_bounding_box_size_rectangle_30_deg() {
+        let size = Size::new(200, 100);
+        let bb_width =
+            60.0f64.to_radians().sin() * size.width.0 + 30.0f64.to_radians().sin() * size.height.0;
+        let bb_height =
+            60.0f64.to_radians().sin() * size.height.0 + 30.0f64.to_radians().sin() * size.width.0;
+        let bb_size = Size::new(bb_width, bb_height);
+        for rotation in &[-150.0, -30.0, 30.0, 150.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(
+                Size,
+                bb_size,
+                bounding_box_offset_and_size(&rotation, &size).1
+            );
         }
+    }
+
+    #[test]
+    fn test_bounding_box_size_rectangle_45_deg() {
+        let size = Size::new(200, 100);
+        let bb_width = 45.0f64.to_radians().sin() * (size.width.0 + size.height.0);
+        let bb_size = Size::new(bb_width, bb_width);
+        for rotation in &[-135.0, -45.0, 45.0, 135.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(
+                Size,
+                bb_size,
+                bounding_box_offset_and_size(&rotation, &size).1
+            );
+        }
+    }
+
+    #[test]
+    fn test_bounding_box_size_rectangle_60_deg() {
+        let size = Size::new(200, 100);
+        let bb_width =
+            30.0f64.to_radians().sin() * size.width.0 + 60.0f64.to_radians().sin() * size.height.0;
+        let bb_height =
+            30.0f64.to_radians().sin() * size.height.0 + 60.0f64.to_radians().sin() * size.width.0;
+        let bb_size = Size::new(bb_width, bb_height);
+        for rotation in &[-120.0, -60.0, 60.0, 120.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(
+                Size,
+                bb_size,
+                bounding_box_offset_and_size(&rotation, &size).1
+            );
+        }
+    }
+
+    #[test]
+    fn test_bounding_box_size_rectangle_90_deg() {
+        let size = Size::new(200, 100);
+        let bb_size = Size::new(100, 200);
+        for rotation in &[-90.0, 90.0] {
+            println!("rotation = {}", rotation);
+            let rotation = Rotation::from(*rotation);
+            assert_approx_eq!(
+                Size,
+                bb_size,
+                bounding_box_offset_and_size(&rotation, &size).1
+            );
+        }
+    }
+
+    #[test]
+    fn test_bounding_box_position_rectangle_30_deg() {
+        let size = Size::new(200, 100);
+        let bb_width =
+            30.0f64.to_radians().sin() * size.width.0 + 60.0f64.to_radians().sin() * size.height.0;
+        let bb_height =
+            30.0f64.to_radians().sin() * size.height.0 + 60.0f64.to_radians().sin() * size.width.0;
+
+        let h30 = 30.0f64.to_radians().cos() * size.height.0;
+        let h60 = 60.0f64.to_radians().cos() * size.height.0;
+        let w30 = 30.0f64.to_radians().cos() * size.width.0;
+        let w60 = 60.0f64.to_radians().cos() * size.width.0;
+
+        test_position(size, -150.0, Position::new(w30, 0));
+        test_position(size, -120.0, Position::new(w60, 0));
+        test_position(size, -60.0, Position::new(0, h60));
+        test_position(size, -30.0, Position::new(0, h30));
+        test_position(size, 30.0, Position::new(h60, bb_width));
+        test_position(size, 60.0, Position::new(h30, bb_height));
+        test_position(size, 120.0, Position::new(bb_width, bb_height - h60));
+        test_position(size, 150.0, Position::new(bb_height, bb_width - h30));
+    }
+
+    #[test]
+    fn test_bounding_box_position_rectangle_45_deg() {
+        let size = Size::new(200, 100);
+        let bb_width = 45.0f64.to_radians().sin() * (size.width.0 + size.height.0);
+
+        test_position(size, -135.0, Position::new(2.0 * bb_width / 3.0, 0));
+        test_position(size, -45.0, Position::new(0, bb_width / 3.0));
+        test_position(size, 45.0, Position::new(bb_width / 3.0, bb_width));
+        test_position(size, 135.0, Position::new(bb_width, 2.0 * bb_width / 3.0));
+    }
+
+    #[test]
+    fn test_bounding_box_position_rectangle_90_deg() {
+        let size = Size::new(200, 100);
+        test_position(size, -180.0, Position::new(200, 0));
+        test_position(size, -90.0, Position::new(0, 0));
+        test_position(size, 0.0, Position::new(0, 100));
+        test_position(size, 90.0, Position::new(100, 200));
+        test_position(size, 180.0, Position::new(200, 0));
     }
 }